HSEARCH-3316 Move DocumentRefreshStrategy to API
So that we can expose hooks to allow users to customize it in the
mapper.
